Separate installing and getting a function
