+++ node/function.go
@@ -1,14 +1,9 @@
 package node
 
-import (
-	"github.com/blocklessnetworking/b7s/models/blockless"
-)
-
 // FStore provides retrieval of function manifest.
 type FStore interface {
-	// Get retrieves a function manifest based on the address or CID. `useCached` boolean
-	// determines if function manifest should be refetched or previously cached data can be returned.
-	Get(address string, cid string, useCached bool) (*blockless.FunctionManifest, error)
+	// Install will install a function based on the address and CID.
+	Install(address string, cid string) error
 
 	// Installed returns info if the function is installed or not.
 	Installed(cid string) (bool, error)
